[Feature #161250122] rename product model
