+++ application/api/v1/models/product_model.py
